Created FreeListAllocator, allows for allocation and deallocation in any order. PointerUtil::ConvertToBinaryString now functions with any object, not just unsigned integers. Fixed crucial bug in pointer alignment arithmetic where a type conversion was missing, causing 32 bit alignments to act real funky. MemoryManager now has properly setup administration. MemoryManager now allows you to create allocators which are placed in the memory we've pre-allocated in the MemoryManager's administration memory. MemoryManager now also manages all the memory that is left in the memory pool. Fixed compiler warning in LinearAllocator because of a missing type conversion. Added macros for easily sending events to the HeapInspectorClient about freeing (deallocating) memory.

+++ blowbox/core/memory/linear_allocator.cc
@@ -35,7 +35,7 @@ namespace blowbox
 
 			uintptr_t aligned_address = (uintptr_t)marker_ + adjustment;
 			
-			marker_ = PointerUtil::Add(marker_, size + adjustment);
+			marker_ = PointerUtil::Add(marker_, (uint32_t)size + adjustment);
 			used_memory_ += adjustment + size;
 			num_allocations_++;
